requirements.txt
django>=1.7,<1.8
django-tastypie
ijson
lxml
defusedxml
